misc: Implement Jekyll plugin to load remote markdown files.
Documentation pages could make use of GitHub README.md files from the
projects GitHub repositories directly. This avoids having duplicate
documentation information.

To achieve this goal, I wrote a Jekyll plugin which is able to load a
markdown file from a given remote URI.

To use it, it is as simple as adding this tag, somewhere in a markdown
or html file:

    {% remote_markdown http://foo.org/bar.md %}

And all the content of the remote markdown file will be processed by
Jekyll when rendering the page.