build: Add Makefile with publish target.
GitHub pages does not allow Jekyll plugins except for some of them.
Thus, to publish the website, I follow the idea of Randy Morris (see
[1]).

So I created a 'source' branch on which Jekyll sources of the website
will reside, and the 'master' branch will contain only the _site/
directory where Jekyll generates the website.

To simplify the publishing process, ie updating the master branch with
only the _site/, add a Makefile with a 'publish' target.

[1]: https://github.com/randymorris/randymorris.github.com/blob/source/README.md